Added `strictLatest` and `keepUnreleased` options

+++ index.js
@@ -11,13 +11,16 @@ const getFormattedDate = () => {
 
 class KeepAChangelog extends Plugin {
   init() {
-    const { filename } = this.options;
+    const { filename, strictLatest, keepUnreleased } = this.options;
 
     this.filename = filename || 'CHANGELOG.md';
+    this.strictLatest = strictLatest === undefined ? true : Boolean(strictLatest);
+    this.keepUnreleased = keepUnreleased === undefined ? false : Boolean(keepUnreleased);
+
     this.changelogPath = path.resolve(this.filename);
     this.changelogContent = fs.readFileSync(this.changelogPath, 'utf-8');
     this.unreleasedTitleRaw = 'Unreleased';
-    this.unreleasedTitle = `\n\n## [${this.unreleasedTitleRaw}]\n\n`;
+    this.unreleasedTitle = `\n\n## [${this.unreleasedTitleRaw}]\n`;
 
     const hasUnreleasedSection = this.changelogContent.includes(this.unreleasedTitle);
     if (!hasUnreleasedSection) {
@@ -29,23 +32,27 @@ class KeepAChangelog extends Plugin {
     const { changelog } = this.getContext();
     if (changelog) return changelog;
 
-    const { filename } = this.options;
-    const previousReleaseTitle = `\n\n## [${latestVersion}]`;
-    const hasPreviouReleaseSection = this.changelogContent.includes(previousReleaseTitle);
-    const startIndex = this.changelogContent.indexOf(this.unreleasedTitle) + this.unreleasedTitle.length;
-    const endIndex = this.changelogContent.indexOf(previousReleaseTitle);
-    const changelogContent = this.changelogContent.substring(startIndex, endIndex).trim();
+    const { filename, strictLatest } = this;
 
-    if (!hasPreviouReleaseSection) {
+    const previousReleaseTitle = strictLatest ? `## [${latestVersion}]` : `## [`;
+    const hasPreviouReleaseSection = this.changelogContent.includes(previousReleaseTitle);
+
+    if (strictLatest && !hasPreviouReleaseSection) {
       throw Error(`Missing section for previous release ("${latestVersion}") in ${filename}.`);
     }
 
+    const startIndex = this.changelogContent.indexOf(this.unreleasedTitle) + this.unreleasedTitle.length;
+    let endIndex = this.changelogContent.indexOf(previousReleaseTitle, startIndex);
+    if (!strictLatest && endIndex === -1) {
+      endIndex = this.changelogContent.length
+    }
+
+    const changelogContent = this.changelogContent.substring(startIndex, endIndex).trim();
     if (!changelogContent) {
       throw Error(`There are no entries under "${this.unreleasedTitleRaw}" section in ${filename}.`);
     }
 
     this.setContext({ changelog: changelogContent });
-
     return changelogContent;
   }
 
@@ -54,11 +61,12 @@ class KeepAChangelog extends Plugin {
   }
 
   beforeRelease() {
+    const { keepUnreleased } = this
     const { isDryRun } = this.global;
-    if (isDryRun) return;
+    if (isDryRun || keepUnreleased) return;
     const { version } = this.getContext();
     const formattedDate = getFormattedDate();
-    const releaseTitle = `\n\n## [${version}] - ${formattedDate}\n\n`;
+    const releaseTitle = `\n\n## [${version}] - ${formattedDate}\n`;
     const changelog = this.changelogContent.replace(this.unreleasedTitle, releaseTitle);  
     fs.writeFileSync(this.changelogPath, changelog);
   }
